adds 2 new badge templates
badge-sm is a small badge for smaller badge holders. It should come out to be 3.5"x2.25"
badge-tall is a taller badge template that should come out to 3.25"x4.5"
I don't have a printer, so it would be good if someone who does could print these out to make sure they are the right size.
[needs testing:295]